Move types from CustomerDashboard component to types file
