+++ src/components/CustomerDashboard/CustomerDashboard.tsx
@@ -11,24 +11,14 @@ import { useParams } from 'react-router';
 import { getOneCustomer } from '../../apiCalls';
 import { NavigationBarProps } from '../NavigationBar/NavigationBar';
 import ErrorPage from '../ErrorPage/ErrorPage';
-
-type Vendor = {
-  email: string;
-  first_name: string;
-  id: number;
-  last_name: string;
-  location: string;
-  market: number;
-  vendor_name: string;
-};
-
-export type CustomerDashboardProps = {
-  selectedZipcode: string;
-  selectedRadius: string;
-  allVendors: Vendor[];
-  isVendor: boolean;
-  currentUserId: string;
-};
+import {
+  Vendor,
+  User,
+  CustomerDashboardProps,
+  selectedVendorItem,
+  CurrentCustomer,
+  CustomerViewItemCardProps,
+} from '../../types';
 
 const CustomerDash = ({
   selectedZipcode,
@@ -49,9 +39,6 @@ const CustomerDash = ({
       .catch(error => setCustomerDashOneCustomerError(error.message));
   }, []);
 
-  type User = {
-    first_name?: string;
-  };
   const [currentUserObj, setCurrentUserObj] = useState<User>({});
 
   useEffect(() => {
@@ -233,40 +220,6 @@ const CustomerDash = ({
       </section>
     </div>
   );
-};
-
-type selectedVendorItem = {
-  id: number;
-  item_name: string;
-  vendor: number;
-  price: string;
-  size: string;
-  quantity: number;
-  availability: boolean;
-  description: string;
-  image: string;
-  date_created: string;
-  updated_at: string;
-};
-
-type CurrentCustomer = {
-  first_name: string;
-  email: string;
-};
-
-type CustomerViewItemCardProps = {
-  availability: boolean;
-  id: number;
-  item_name: string;
-  price: string;
-  size: string;
-  item_quantity: number;
-  description: string;
-  selectedVendorObject: Vendor | null;
-  onUpdateQuantity: (itemId: number, updatedQuantity: number) => void;
-  currentUser?: CurrentCustomer;
-  image: string;
-  setMessage: Function;
 };
 
 const CustomerViewItemCard = ({
@@ -458,24 +411,7 @@ const CustomerViewItemCard = ({
             <span className="close" onClick={closeModal}>
               &times;
             </span>
-            <h2 className="invoice-header">
-              Order Confirmation{' '}
-              {/* <svg
-                xmlns="http://www.w3.org/2000/svg"
-                fill="none"
-                viewBox="0 0 24 24"
-                strokeWidth="1.5"
-                stroke="currentColor"
-                className="printer-icon"
-                onClick={handlePrint}
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  d="M6.72 13.829c-.24.03-.48.062-.72.096m.72-.096a42.415 42.415 0 0110.56 0m-10.56 0L6.34 18m10.94-4.171c.24.03.48.062.72.096m-.72-.096L17.66 18m0 0l.229 2.523a1.125 1.125 0 01-1.12 1.227H7.231c-.662 0-1.18-.568-1.12-1.227L6.34 18m11.318 0h1.091A2.25 2.25 0 0021 15.75V9.456c0-1.081-.768-2.015-1.837-2.175a48.055 48.055 0 00-1.913-.247M6.34 18H5.25A2.25 2.25 0 013 15.75V9.456c0-1.081.768-2.015 1.837-2.175a48.041 48.041 0 011.913-.247m10.5 0a48.536 48.536 0 00-10.5 0m10.5 0V3.375c0-.621-.504-1.125-1.125-1.125h-8.25c-.621 0-1.125.504-1.125 1.125v3.659M18 10.5h.008v.008H18V10.5zm-3 0h.008v.008H15V10.5z"
-                />
-              </svg> */}
-            </h2>
+            <h2 className="invoice-header">Order Confirmation </h2>
             <p className="invoice-invoice-num">
               <strong>INVOICE #:</strong> 12
             </p>
