+++ src/types.tsx
@@ -1,0 +1,55 @@
+export type Vendor = {
+  email: string;
+  first_name: string;
+  id: number;
+  last_name: string;
+  location: string;
+  market: number;
+  vendor_name: string;
+};
+
+export type selectedVendorItem = {
+  id: number;
+  item_name: string;
+  vendor: number;
+  price: string;
+  size: string;
+  quantity: number;
+  availability: boolean;
+  description: string;
+  image: string;
+  date_created: string;
+  updated_at: string;
+};
+
+export type CustomerDashboardProps = {
+  selectedZipcode: string;
+  selectedRadius: string;
+  allVendors: Vendor[];
+  isVendor: boolean;
+  currentUserId: string;
+};
+
+export type CustomerViewItemCardProps = {
+  availability: boolean;
+  id: number;
+  item_name: string;
+  price: string;
+  size: string;
+  item_quantity: number;
+  description: string;
+  selectedVendorObject: Vendor | null;
+  onUpdateQuantity: (itemId: number, updatedQuantity: number) => void;
+  currentUser?: CurrentCustomer;
+  image: string;
+  setMessage: Function;
+};
+
+export type User = {
+  first_name?: string;
+};
+
+export type CurrentCustomer = {
+  first_name: string;
+  email: string;
+};
